millis() only returns an int

+++ beep.c
@@ -7,7 +7,7 @@ void crap_beep(unsigned int pin, unsigned int note, unsigned int dur)
 {
     DDRB |= (1<<SND);
     
-    unsigned long ts = millis();
+    unsigned int ts = millis();
     while (millis() < ts + dur)
     {
         PORTB |= 1 << pin;    // HIGH
